New minor version 4.01
4.01: March 2014
- More info for RAC Cross Instance Parallel Execution: Many sections now
show a GLOBAL aggregate info in addition to instance-specific data

- The Parallel Execution Server Set detection and ASSUMED_DEGREE info
now makes use of the undocumented PX_STEP_ID and PX_STEPS_ARG info (bit
mask part of the PX_FLAGS column) on 11.2.0.2+

- Since version 4.0 added from 11.2.0.2 on the PX *MAX* DOP in the "SQL
statement execution ASH Summary" based on new PX_FLAGS column of ASH it
makes sense to add a PX *MIN* DOP in the summary to see at one glance if
different DOPs were used or not

- The "Active DOPs" column in the "Activity Timeline based on ASH" was
extended/modified: The number in parantheses is no longer the simple
count of samples but the Average Active Sessions (AAS) per DFO / bucket.
From 11.2.0.2 it now shows also the DOP of the DFO in brackets, so the
output could look now like this:

1[16] (14.5)

which means DFO 1 at a DOP of 16 had an AAS value of 14.5 for this time
bucket. If there are multiple DFOs active in the time bucket, they are
separated by commas:

1[16] (3.5),2[4] (1.5)

which means DFO 1 at a DOP of 16 had an AAS value of 3.5 and DFO 2 at a
DOP of 4 had an AAS value of 1.5 for this time bucket

- The "Real-Time SQL Monitoring Execution Summary" section now shows a
PX *MIN* DOP information, too, based on the monitoring information.
Note that from version 11.2 on there is a column PX_MAXDOP, but no
PX_MINDOP column. So from version 11.2 on for PX_MAXDOP the column value
will be used.
For pre-11.2 versions the PX_MAXDOP is calculated using an analytic
function on the monitoring data
In the same way the PX_MINDOP information is now calculated in all
versions that support monitoring

Please note that both PX_MAXDOP and PX_MINDOP are unreliable in 11.1
when the execution plan consists of multiple DFOs